backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
pymongo==4.10.1
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.5
Pillow==11.1.0
